Chatbot on ollam/requirement.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
langchain-objectbox